% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closest_stations.R
\name{get_closest_stations}
\alias{get_closest_stations}
\title{Get closest SMHI stations with available data}
\usage{
get_closest_stations(
  points,
  parameter_key,
  period_name,
  max_radius = 100,
  n = 1
)
}
\arguments{
\item{points}{An \code{sf} object (of type \code{POINT}) representing one or more spatial locations.}

\item{parameter_key}{A character vector of SMHI parameter IDs to search for (e.g. \code{"1"} for temperature).}

\item{period_name}{A character string indicating the period for which data is requested (e.g., \code{"latest-month"}).
Options include "latest-day", "latest-hour", "latest-months" and "corrected-archive".}

\item{max_radius}{Numeric. Maximum distance (in kilometers) to search for stations. Default is \code{100}.}

\item{n}{Integer. Number of closest stations to retrieve per point and parameter. Default is \code{1}.}
}
\value{
A tibble with information about the closest stations for each spatial point and parameter,
including columns for \code{parameter_id}, \code{station_id}, and \code{distance.km}, as well as the original coordinates.
}
\description{
This function identifies the closest meteorological stations (from a preloaded \code{stations_smhi} dataset)
to one or more spatial points (\code{sf::POINT}), filtering by available parameters and time periods.
}
\details{
This function filters the internal \code{stations_smhi} dataset to find stations matching the given
\code{parameter_key}, and checks that the stations contain data for the selected \code{period_name} using
\code{get_smhi_period()}. Only stations within the given \code{max_radius} are considered.
}
\examples{
\dontrun{
library(sf)
point <- sf::st_sf(geometry = sf::st_sfc(sf::st_point(c(18.07, 59.33)), crs = 4326))
get_closest_stations(point, parameter_key = "1", period_name = "latest-month")
}
}
\seealso{
\code{\link[=get_smhi_period]{get_smhi_period()}}, \code{\link[=get_smhi_data2]{get_smhi_data2()}}, \code{\link[=get_smhi_closest]{get_smhi_closest()}}
}
