% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smhi_closest.R
\name{get_smhi_closest}
\alias{get_smhi_closest}
\title{Get SMHI meteorological data from the closest stations to spatial points}
\usage{
get_smhi_closest(
  points,
  parameter_key,
  period_name,
  verbose = TRUE,
  n = 1,
  max_radius = 100
)
}
\arguments{
\item{points}{An \code{sf} object (of type \code{POINT}) representing one or more spatial locations.}

\item{parameter_key}{A character vector with the SMHI parameter IDs (e.g. \code{"1"} for temperature).}

\item{period_name}{A character string indicating the period for which the data is requested (e.g., \code{"latest-hour"}).
Options include "latest-day", "latest-hour", "latest-months" and "corrected-archive".}

\item{verbose}{Logical. Whether to display messages during processing. Default is \code{TRUE}.}

\item{n}{Integer. Number of closest stations to retrieve for each point. Default is \code{1}.}

\item{max_radius}{Numeric. Maximum distance (in kilometers) to search for stations. Default is \code{100}.}
}
\value{
A tibble with meteorological data from the closest available stations for each point and parameter.
If no stations are found within the given radius, returns \code{NULL}.
}
\description{
This function identifies the closest meteorological stations to one or more spatial points (\code{sf::POINT})
that have available data for the selected parameters and time period. It then downloads the corresponding
meteorological data using \code{get_smhi_data2()}.
}
\details{
This function first uses \code{get_closest_stations()} to find nearby stations for each point that match the given
\code{parameter_key} and \code{period_name}. Then it downloads and compiles the corresponding data using \code{get_smhi_data2()}.
}
\examples{
\dontrun{
library(sf)
points <- sf::st_sf(geometry = sf::st_sfc(sf::st_point(c(18.07, 59.33)), crs = 4326))
get_smhi_closest(points, parameter_key = "1", period_name = "latest-hour")
}
}
\seealso{
\code{\link[=get_closest_stations]{get_closest_stations()}}, \code{\link[=get_smhi_data2]{get_smhi_data2()}}
}
