% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smhi_data.R
\name{get_smhi_data}
\alias{get_smhi_data}
\title{Get SMHI data for a given station, parameter, and period}
\usage{
get_smhi_data(station_key, parameter_key, period_name, verbose = T)
}
\arguments{
\item{station_key}{A character vector of station identifiers. Check \code{get_smhi_station()} for stations names.}

\item{parameter_key}{A character vector of parameter identifiers. Check \code{get_smhi_parameters()} .}

\item{period_name}{A character string indicating the period for which the data is requested.
Options include "latest-day", "latest-hour", "latest-months" and "corrected-archive".}

\item{verbose}{A logical value indicating whether to print messages about the progress of the data download. Default is TRUE.}
}
\value{
A tibble containing the meteorological data for the specified station, parameter, and period.
}
\description{
This function retrieves meteorological data from the Swedish Meteorological and Hydrological Institute (SMHI)
for a specified station, parameter, and period.
}
\examples{
\dontrun{
# Get the latest hour temperature data for station 22051
get_smhi_data(station_key = "22051", parameter_key = "1", period_name = "latest-hour")
}

}
