% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smhi_data2.R
\name{get_smhi_data2}
\alias{get_smhi_data2}
\title{Get meteorological data for multiple station/parameter pairs}
\usage{
get_smhi_data2(pairs_tbl, period_name, verbose = TRUE)
}
\arguments{
\item{pairs_tbl}{A tibble containing two columns: \code{station_id} and \code{parameter_id}. Check \code{get_smhi_station()} and \code{get_smhi_parameters()} for
stations and parameters information
Each row represents a station/parameter pair for which data will be retrieved.}

\item{period_name}{A character string indicating the period for which the data is requested.
Options include "latest-day", "latest-hour", "latest-months" and "corrected-archive".}

\item{verbose}{A logical value indicating whether to print messages about the progress of the data download. Default is TRUE.}
}
\value{
A tibble containing the meteorological data for the specified station/parameter pairs and period.
}
\description{
This function retrieves meteorological data from the Swedish Meteorological and Hydrological Institute (SMHI) for multiple station/parameter pairs
and returns the data as a tidy tibble.
}
\details{
The function calls the \code{get_smhi_data()} function internally to retrieve data for each station/parameter pair.
}
\examples{
\dontrun{
# Example usage to get data for multiple station/parameter pairs
pairs_tbl <- tibble::tibble(
  station_id = c("22051", "159880"),
  parameter_id = c("1", "21")
)
get_smhi_data2(pairs_tbl, period_name = "latest-hour")
}

}
\seealso{
\code{\link[=get_smhi_station]{get_smhi_station()}}, \code{\link[=get_smhi_parameters]{get_smhi_parameters()}}
}
