% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smhi_parameters.R
\name{get_smhi_parameters}
\alias{get_smhi_parameters}
\title{Get SMHI Parameters Information}
\usage{
get_smhi_parameters(version = "latest")
}
\arguments{
\item{version}{A character string specifying the version of the SMHI API to use (default is "latest").
Chech \verb{https://opendata-download-metobs.smhi.se/metobs/introduction} for available versions.}
}
\value{
A tibble containing the following columns:
\item{parameter_id}{The ID of the parameter.}
\item{title}{The title or name of the parameter.}
\item{summary}{A brief summary or description of the parameter.}
\item{unit}{The unit of measurement for the parameter.}
}
\description{
This function retrieves information about the available parameters from the SMHI (Swedish Meteorological and Hydrological Institute) API.
}
\examples{
get_smhi_parameters()
}
