% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smhi_period.R
\name{get_smhi_period}
\alias{get_smhi_period}
\title{Get available periods for a specific station and parameter}
\usage{
get_smhi_period(station_key, parameter_key, version = "latest")
}
\arguments{
\item{station_key}{A character string representing the station ID. Check \code{get_smhi_station()} for stations names.}

\item{parameter_key}{A character string representing the parameter ID. Check \code{get_smhi_parameters()} .}

\item{version}{A character string indicating the version of the API to use. Default is "latest".
Chech \verb{https://opendata-download-metobs.smhi.se/metobs/introduction} for available versions.}
}
\value{
A character vector of period names representing the available periods for the specified station and parameter.
}
\description{
This function retrieves the available periods for a given station and parameter
from the SMHI API. The periods represent the time spans for which data is available
for the specified station and parameter.
}
\examples{
# Retrieve available periods for a specific station and parameter
periods <- get_smhi_period(station_key = "159880", parameter_key = "1")

}
