% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smhi_station.R
\name{get_smhi_station}
\alias{get_smhi_station}
\title{Get SMHI station information for specified parameters}
\usage{
get_smhi_station(parameter_key, version = "latest", spatial = F)
}
\arguments{
\item{parameter_key}{A character vector containing the parameter IDs (e.g., "1" for temperature) to retrieve station data for.}

\item{version}{A character string indicating the version of the API to use. Default is "latest".
Chech \verb{https://opendata-download-metobs.smhi.se/metobs/introduction} for available versions.}

\item{spatial}{A logical value indicating whether the results should be returned as a spatial object (\code{sf} object).
If \code{TRUE}, the result will be a spatial data frame with coordinates for each station.
Default is \code{FALSE}.}
}
\value{
A data frame (or \code{sf} object if \code{spatial = TRUE}) containing the following columns:
\item{parameter_id}{The parameter ID.}
\item{station_id}{The station ID.}
\item{station_name}{The name of the station.}
\item{longitude}{The longitude of the station.}
\item{latitude}{The latitude of the station.}
\item{height}{The elevation (in metres) of the station.}
\item{owner}{The owner of the station.}
\item{ownerCategory}{The category of the owner.}
\item{active}{Indicates whether the station is active.}
\item{from}{The start date of station data.}
\item{to}{The end date of station data.}
\item{updated}{The date when the station information was last updated.}
}
\description{
This function retrieves information about SMHI stations based on the provided
parameter IDs. It supports the option to return spatial data in the form of
an \code{sf} object with geographic coordinates.
}
\examples{
# Retrieve station information for temperature (parameter_id = "1") from the latest version
stations <- get_smhi_station(parameter_key = c("1"))

# Retrieve station information with spatial data (sf object)
stations_spatial <- get_smhi_station(parameter_key = c("1"), spatial = TRUE)

}
\seealso{
\code{\link[=get_smhi_station]{get_smhi_station()}}, \code{\link[=get_smhi_parameters]{get_smhi_parameters()}}
}
